+++ api/src/utils/dgraphClient.ts
@@ -1,0 +1,4 @@
+import { DgraphClient, DgraphClientStub } from 'dgraph-js-http';
+
+const dgraphClientStub = new DgraphClientStub('http://localhost:8080');
+export const dgraphClient = new DgraphClient(dgraphClientStub);
